refactor(inventory): Phase 7 relocate InternalOutboxEndpoints + Program.cs manifest
- Move Endpoints/InternalOutboxEndpoints.cs to Infrastructure/Outbox/ with
  namespace Inventory.Service.Infrastructure.Outbox; delete now-empty Endpoints/.
- Program.cs using directive swapped to new namespace; manifest shape preserved
  (slice registrations + mappings + MapHealthChecks + RegisterInternalOutboxEndpoints).
- RequireService policy gate on /internal/outbox/failed unchanged.
